Predict useful mutation positions/substitutions
Using results from laboratory trials, this script makes uses of statistical concepts of Akaike's Transformation. A matrix based table is generated. Box cox y transformation is applied on the model and results in predicting which mutation position and substitution have been working for the given molecule. The script is useful specially when multi position substitutions are added in parallel to a given molecule and when the effect of each is difficult to determine individually.