Visualization for Site saturated library mutations
The script takes in a file with just 2 basic inputs - mutations in positions and the responses at those positions. Generates a graph for each position for each of the 20 amino acids. Holistic visualization using color coding makes it easy to quickly see which positions and which amino acids have been tried in the laboratory so far, which substitutions seem to be working for the given position.